Dream Texture: UI Changes
* Changed UI from what was a three-click process to generate a new Stable Diffusion to a one-click process.
* "Dream Texture" text was very wide in the Image Editor & Node Editor menus & hard to access, now in collapsible sidebar
* Organized UI module into folder
* Assigned more appropriate History icon
* Smaller UI tweaks
