+++ operators/dream_texture.py
@@ -30,10 +30,6 @@ class DreamTexture(bpy.types.Operator):
     bl_label = "Dream Texture"
     bl_description = "Generate a texture with AI"
     bl_options = {'REGISTER'}
-
-    @classmethod
-    def poll(self, context):
-        return True
     
     def invoke(self, context, event):
         weights_installed = os.path.exists(WEIGHTS_PATH)
@@ -41,66 +37,7 @@ class DreamTexture(bpy.types.Operator):
             self.report({'ERROR'}, "Please complete setup in the preferences window.")
             return {"FINISHED"}
         else:
-            return context.window_manager.invoke_props_dialog(self)
-
-    def draw(self, context):
-        layout = self.layout
-        
-        scene = context.scene
-        
-        prompt_box = layout.box()
-        prompt_box_heading = prompt_box.row()
-        prompt_box_heading.label(text="Prompt")
-        prompt_box_heading.prop(scene.dream_textures_prompt, "prompt_structure")
-        structure = next(x for x in prompt_structures if x.id == scene.dream_textures_prompt.prompt_structure)
-        for segment in structure.structure:
-            segment_row = prompt_box.row()
-            enum_prop = 'prompt_structure_token_' + segment.id + '_enum'
-            is_custom = getattr(scene.dream_textures_prompt, enum_prop) == 'custom'
-            if is_custom:
-                segment_row.prop(scene.dream_textures_prompt, 'prompt_structure_token_' + segment.id)
-            segment_row.prop(scene.dream_textures_prompt, enum_prop, icon_only=is_custom)
-        
-        size_box = layout.box()
-        size_box.label(text="Configuration")
-        size_box.prop(scene.dream_textures_prompt, "width")
-        size_box.prop(scene.dream_textures_prompt, "height")
-        size_box.prop(scene.dream_textures_prompt, "seamless")
-        
-        for area in context.screen.areas:
-            if area.type == 'IMAGE_EDITOR':
-                if area.spaces.active.image is not None and image_has_alpha(area.spaces.active.image):
-                    inpainting_box = layout.box()
-                    inpainting_heading = inpainting_box.row()
-                    inpainting_heading.prop(scene.dream_textures_prompt, "use_inpainting")
-                    inpainting_heading.label(text="Inpaint Open Image")
-                    break
-
-        if not scene.dream_textures_prompt.use_inpainting:
-            init_img_box = layout.box()
-            init_img_heading = init_img_box.row()
-            init_img_heading.prop(scene.dream_textures_prompt, "use_init_img")
-            init_img_heading.label(text="Init Image")
-            if scene.dream_textures_prompt.use_init_img:
-                init_img_box.template_ID(context.scene, "init_img", open="image.open")
-                init_img_box.prop(scene.dream_textures_prompt, "strength")
-                init_img_box.prop(scene.dream_textures_prompt, "fit")
-
-        advanced_box = layout.box()
-        advanced_box_heading = advanced_box.row()
-        advanced_box_heading.prop(scene.dream_textures_prompt, "show_advanced", icon="DOWNARROW_HLT" if scene.dream_textures_prompt.show_advanced else "RIGHTARROW_THIN", emboss=False, icon_only=True)
-        advanced_box_heading.label(text="Advanced Configuration")
-        if scene.dream_textures_prompt.show_advanced:
-            advanced_box.prop(scene.dream_textures_prompt, "full_precision")
-            advanced_box.prop(scene.dream_textures_prompt, "seed")
-            # advanced_box.prop(self, "iterations") # Disabled until supported by the addon.
-            advanced_box.prop(scene.dream_textures_prompt, "steps")
-            advanced_box.prop(scene.dream_textures_prompt, "cfgscale")
-            advanced_box.prop(scene.dream_textures_prompt, "sampler")
-            advanced_box.prop(scene.dream_textures_prompt, "show_steps")
-
-    def cancel(self, context):
-        pass
+            return self.execute(context)
 
     async def dream_texture(self, context):
         history_entry = context.preferences.addons[StableDiffusionPreferences.bl_idname].preferences.history.add()
@@ -148,7 +85,7 @@ class DreamTexture(bpy.types.Operator):
                 if last_data_block is not None:
                     bpy.data.images.remove(last_data_block)
                     last_data_block = None
-                image = pil_to_image(image, name=f"{seed}")
+                image = pil_to_image(image, name=f"{generated_prompt}")
                 if node_tree is not None:
                     nodes = node_tree.nodes
                     texture_node = nodes.new("ShaderNodeTexImage")
@@ -207,7 +144,7 @@ class DreamTexture(bpy.types.Operator):
                 init_img_path = save_temp_image(init_img)
 
             generator.prompt2image(
-                # prompt string (no default)
+                # prompt string (no default)                
                 prompt=generated_prompt,
                 # iterations (1); image count=iterations
                 iterations=scene.dream_textures_prompt.iterations,
